Divide camera into two classes
