+++ src/engine/demo/DitheringStage.h
@@ -46,7 +46,7 @@ public:
     }
 
     void renderUI() override;
-    void renderContent(Camera camera, double dt) override;
+    void renderContent(FreeCamera camera, double dt) override;
 };
 
 #endif // DITHERINGSTAGE_H
