+++ src/engine/demo/MeshStage.cpp
@@ -109,7 +109,7 @@ void MeshStage::renderUI() {
     UIStage::renderUI();
 }
 
-void MeshStage::renderContent(Camera camera, double dt) {
+void MeshStage::renderContent(FreeCamera camera, double dt) {
     Mat4 view = camera.getViewMatrix();
     Mat4 mm = Mat4::identity();
     mm = Mat4::scale(0.75f) * mm;
