+++ src/engine/demo/MeshStage.h
@@ -32,7 +32,7 @@ public:
     }
 
     void renderUI() override;
-    void renderContent(Camera camera, double dt) override;
+    void renderContent(FreeCamera camera, double dt) override;
 };
 
 #endif // MESHSTAGE_H
