+++ src/engine/demo/ShadingStage.h
@@ -36,7 +36,7 @@ public:
     }
 
     void renderUI() override;
-    void renderContent(Camera camera, double dt) override;
+    void renderContent(FreeCamera camera, double dt) override;
 };
 
 #endif // SHADINGSTAGE_H
