+++ src/engine/demo/TextureStage.cpp
@@ -58,7 +58,7 @@ void TextureStage::renderUI() {
     UIStage::renderUI();
 }
 
-void TextureStage::renderContent(Camera camera, double dt) {
+void TextureStage::renderContent(FreeCamera camera, double dt) {
     Mat4 view = camera.getViewMatrix();
     Mat4 mm = Mat4::identity();
     mm = Mat4::scale(0.75f) * mm;
