+++ src/engine/demo/TextureStage.h
@@ -40,7 +40,7 @@ public:
     }
 
     void renderUI() override;
-    void renderContent(Camera camera, double dt) override;
+    void renderContent(FreeCamera camera, double dt) override;
 };
 
 #endif // TEXTURESTAGE_H
