+++ src/engine/gui/UIStage.h
@@ -2,7 +2,7 @@
 #define UISTAGE_H
 
 #include <memory>
-#include <core/Camera.h>
+#include <core/FreeCamera.h>
 #include "UIComponent.h"
 #include "UIRenderer.h"
 
@@ -15,7 +15,7 @@ public:
     virtual void click(const double &x, const double &y);
     virtual void cursor(const double &x, const double &y);
     virtual void renderUI();
-    virtual void renderContent(Camera camera, double dt) = 0;
+    virtual void renderContent(FreeCamera camera, double dt) = 0;
 };
